+++ NumberTrivia/Domain/Repositories/NumberTriviaRepository.swift
@@ -9,5 +9,5 @@ import Foundation
 
 protocol NumberTriviaRepository {
   func getRandomNumberTrivia(response: @escaping NumberTriviaResponse)
-  func getConcreteNumberTrivia(number: Int, response: @escaping NumberTriviaResponse)
+  func getConcreteNumberTrivia(number: Double, response: @escaping NumberTriviaResponse)
 }
